table scale finaly, table filter fixed, table recoloring finaly, parser fixed

+++ db/init_db.py
@@ -10,7 +10,7 @@ def init_db():
       industry TEXT DEFAULT "",
       sector TEXT DEFAULT "",
       title TEXT,
-      ebitda INTEGER,
+      ebitda REAL,
       net_profit_margin REAL, 
       p_e REAL,
       p_s REAL,
